requirements.txt
astroid==2.2.5
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-rest-auth==0.9.5
djangorestframework==3.10.1
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
oauthlib==3.0.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==3.2.0
social-auth-storage-mongoengine==1.0.1
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
